requirements.txt
Django==4.2.2
phonenumbers==8.13.27
django-phonenumber-field==7.3.0
djangorestframework==3.14.0
Pillow
django-celery-results==2.5.1
django_filter==23.5
redis==5.0.1
kavenegar==1.1.2
python-decouple==3.8
djangorestframework-simplejwt==5.3.1
mysqlclient==2.2.4
gunicorn